blog/requirements.txt
blinker==1.7.0
click==8.1.7
flask==3.0.3
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.1.0
importlib-resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
SQLAlchemy==2.0.29
typing-extensions==4.11.0
werkzeug==3.0.2
zipp==3.18.1
